Added 2 Headers files for enum : PokemonChoice & PokemonType
