+++ PokemonChoice.hpp
@@ -1,0 +1,8 @@
+// Define an enum for Pokemon choices
+enum class PokemonChoice
+{
+    Charmander = 1,
+    Bulbasaur,
+    Squirtle,
+    Pikachu // Default choice
+};
