+++ PokemonType.hpp
@@ -1,0 +1,9 @@
+// Define an enum for Pokemon types
+enum class PokemonType
+{
+    Fire,
+    Grass,
+    Water,
+    Electric,
+    Normal // Added for the default constructor
+};
